+++ packages/prototypes/base/src/separator/index.ts
@@ -1,6 +1,7 @@
 export { asSeparatorRoot } from './root.proto';
 export { default, default as separatorRoot } from './root.proto';
 export type {
+  SeparatorOrientation,
   SeparatorRootProps,
   SeparatorRootExposes,
   SeparatorRootStateHandles,